split r2k file
print cr1 as 0
