+++ libr/io/p/r2k.mk
@@ -1,4 +1,10 @@
-OBJ_R2K=io_r2k.o
+OBJ_R2K = io_r2k.o
+ifeq ($(OSTYPE),$(filter $(OSTYPE),gnulinux android))
+OBJ_R2K += io_r2k_linux.o
+endif
+ifeq (${OSTYPE},windows)
+OBJ_R2K += io_r2k_windows.o
+endif
 
 STATIC_OBJ+=${OBJ_R2K}
 TARGET_R2K=io_r2k.${EXT_SO}
